+++ src/char.h
@@ -113,4 +113,7 @@ static inline bool fly_question(char c)
 	return c == 0x3F ? true : false;
 }
 
+#define FLY_SLASH						'/'
+#define FLY_QUESTION					'?'
+
 #endif