+++ src/v2.c
@@ -105,7 +105,6 @@ static inline uint32_t fly_hv2_length_from_frame_header(fly_hv2_frame_header_t *
 		*(__l+1) = (uint8_t) (*__fh)[1];
 		*(__l+2) = (uint8_t) (*__fh)[0];
 #endif
-//		length |= (uint32_t) (*__fh)[2] << 0;
 	}
 	return length;
 }
@@ -531,7 +530,7 @@ int fly_receive_v2(fly_sock_t fd, fly_connect_t *connect)
 		struct fly_err *__err;
 		__err = fly_err_init(
 			connect->pool, 0, FLY_ERR_ERR,
-			"http2 request receive no buffer chain error in receiving request . (%s: %s)",
+			"http2 request receive no buffer chain error in receiving request . (%s: %d)",
 			__FILE__, __LINE__
 		);
 		fly_error_error(__err);
@@ -1727,7 +1726,7 @@ int fly_hv2_request_event_handler(fly_event_t *event)
 			struct fly_err *__err;
 			__err = fly_event_err_init(
 				event, errno, FLY_ERR_CRIT,
-				"time handler is broken. (%s: %s)",
+				"time handler is broken. (%s: %d)",
 				__FILE__, __LINE__
 			);
 			fly_event_error_add(event, __err);
@@ -2063,7 +2062,7 @@ emergency:
 			event,
 			errno,
 			FLY_ERR_ERR,
-			"HTTP2 received emergency error. (%s: %s)",
+			"HTTP2 received emergency error. (%s: %d)",
 			__FILE__, __LINE__
 		);
 		fly_event_error_add(event, __err);
@@ -2591,7 +2590,7 @@ send:
 							state->pool,
 							errno,
 							FLY_ERR_ERR,
-							"send frame error. (%s: %s)",
+							"send frame error. (%s: %d)",
 							__FILE__, __LINE__
 						);
 						fly_error_error(__err);
@@ -2651,7 +2650,7 @@ send:
 							state->pool,
 							errno,
 							FLY_ERR_ERR,
-							"send frame error. (%s: %s)",
+							"send frame error. (%s: %d)",
 							__FILE__, __LINE__
 						);
 						fly_error_error(__err);
@@ -2716,7 +2715,7 @@ send:
 							state->pool,
 							errno,
 							FLY_ERR_ERR,
-							"send frame error. (%s: %s)",
+							"send frame error. (%s: %d)",
 							__FILE__, __LINE__
 						);
 						fly_error_error(__err);
@@ -2784,7 +2783,7 @@ send:
 							state->pool,
 							errno,
 							FLY_ERR_ERR,
-							"send frame error. (%s: %s)",
+							"send frame error. (%s: %d)",
 							__FILE__, __LINE__
 						);
 						fly_error_error(__err);
@@ -3271,7 +3270,7 @@ __fly_static int __fly_send_frame(struct fly_hv2_send_frame *frame)
 					frame->pool,
 					errno,
 					FLY_ERR_ERR,
-					"send frame error. (%s: %s)",
+					"send frame error. (%s: %d)",
 					__FILE__, __LINE__
 				);
 				fly_error_error(__err);
@@ -4587,6 +4586,8 @@ int fly_hv2_request_line_from_header(fly_request_t *req)
 						printf("\tRequest uri: %s\n", *__p);
 #endif
 						fly_uri_set(req, __c->value, __c->value_len);
+//						if (fly_query_parse_from_uri(req, &req->request_line->uri) == -1)
+//							return -1;
 					}else{
 						FLY_NOT_COME_HERE
 					}
@@ -4760,7 +4761,11 @@ __response:
 	}
 
 	/* defined handler */
-	response = route->function(request, route->data);
+	if (fly_path_param_count(route->path_param) > 0){
+		if (fly_parse_path_params_from_request(request, route) == -1)
+			goto response_500;
+	}
+	response = route->function(request, route, route->data);
 	if (response == NULL)
 		goto response_500;
 	fly_response_header_init(response, request);